==> scripts/01_libraries_setup.Rmd <==
---
title: "Multi-Omics Analysis of Multiple Sclerosis"
author: 
  - "Rofida Gamal (Original Instructor)"
  - "Mohamed Hussein (Code Structuring,Analysis Execution,GitHub Publication)"
date: "`r Sys.Date()`"
output: html_document
---

---

# Task 1: Load Required Libraries & Setup Environment

This task prepares the R environment by loading all required libraries for RNA-seq processing, methylation analysis, statistical modeling, pathway enrichment, and network inference.

---
## load Required Libraries
```{r, message=FALSE, warning=FALSE}
# RNA-seq analysis
library(DESeq2)
library(edgeR)
library(vsn)

# General data wrangling
library(tidyverse)
library(dplyr)
library(genefilter)

# Statistical testing
library(multtest)
library(DescTools)

# Visualization and PCA
library(ggfortify)
library(factoextra)

# Methylation analysis
library(minfi)
library(IlluminaHumanMethylation450kanno.ilmn12.hg19)

# Pathway and enrichment
library(clusterProfiler)
library(org.Hs.eg.db)
library(AnnotationDbi)

# Network inference
library(GeneNet)

# Optional: Check session info to confirm everything loaded
sessionInfo()
```
